Added grouping for hosts and bootfiles.
The bootfiles can now be categorized into 'Bootfiles' and 'Post-install scripts'.
There are also filters in place helping you select the correct type of bootfile
or script when assigning it to the host.

It is also now possible to create host-groups. this allows us to group hosts
togetherm and it makes it possible to have a smaller table showing host status.